+++ app/src/main/java/woowacourse/movie/domain/price/DiscountPolicy.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain
+package woowacourse.movie.domain.price
 
 interface DiscountPolicy {
     fun discount(price: Int): Int
