+++ app/src/main/java/woowacourse/movie/domain/price/PricePolicy.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain
+package woowacourse.movie.domain.price
 
 interface PricePolicy {
